+++ src/application-insights/azext_applicationinsights/util.py
@@ -6,17 +6,19 @@
 from datetime import datetime
 import dateutil.parser  # pylint: disable=import-error
 from msrestazure.tools import is_valid_resource_id, parse_resource_id
-from azext_applicationinsights._client_factory import cf_components, applicationinsights_mgmt_plane_client
+from azext_applicationinsights._client_factory import applicationinsights_mgmt_plane_client
 
 
 def get_id_from_azure_resource(cli_ctx, app, resource_group=None):
     if is_valid_resource_id(app):
         parsed = parse_resource_id(app)
         resource_group, name, subscription = parsed["resource_group"], parsed["name"], parsed["subscription"]
-        client = applicationinsights_mgmt_plane_client(cli_ctx, subscription_id=subscription).components
+        client = applicationinsights_mgmt_plane_client(cli_ctx, subscription_id=subscription,
+                                                       api_version='2015-05-01').components
         return client.get(resource_group, name).app_id
     if resource_group:
-        return cf_components(cli_ctx, None).get(resource_group, app).app_id
+        client = applicationinsights_mgmt_plane_client(cli_ctx, api_version='2015-05-01').components
+        return client.get(resource_group, app).app_id
     return app
 
 
@@ -63,12 +65,12 @@ def get_linked_properties(cli_ctx, app, resource_group, read_properties=None, wr
 
     if isinstance(read_properties, list):
         propLen = len(read_properties)
-        linked_read_properties = ['{}/{}'.format(tmpl, roles[read_properties[i]]) for i in range(propLen)]
+        linked_read_properties = ['{}/{}'.format(tmpl, roles[read_properties[i]]) for i in range(propLen) if read_properties[i]]  # pylint: disable=line-too-long
     else:
         linked_read_properties = ['{}/{}'.format(tmpl, roles[read_properties])]
     if isinstance(write_properties, list):
         propLen = len(write_properties)
-        linked_write_properties = ['{}/{}'.format(tmpl, roles[write_properties[i]]) for i in range(propLen)]
+        linked_write_properties = ['{}/{}'.format(tmpl, roles[write_properties[i]]) for i in range(propLen) if write_properties[i]]  # pylint: disable=line-too-long
     else:
         linked_write_properties = ['{}/{}'.format(tmpl, roles[write_properties])]
     return linked_read_properties, linked_write_properties